fix: keep design image route exports valid
