+++ apps/web/src/app/api/design-docs/generate-image/route.ts
@@ -1,6 +1,10 @@
 import { NextRequest, NextResponse } from "next/server";
 
 import { requireAdmin } from "@/lib/server/auth";
+import {
+  type GeminiModelVariant,
+  resolveGeminiImageModel,
+} from "@/lib/server/design-docs/gemini-image-model";
 
 /* ------------------------------------------------------------------ */
 /*  Design Docs — AI Image Generation API                              */
@@ -9,18 +13,6 @@ import { requireAdmin } from "@/lib/server/auth";
 
 const GEMINI_API_KEY = process.env.GEMINI_API_KEY ?? "";
 const OPENAI_API_KEY = process.env.OPENAI_API_KEY ?? "";
-
-const DEFAULT_GEMINI_IMAGE_MODELS = {
-  flash: "gemini-3.1-flash-image-preview",
-  pro: "gemini-3-pro-image-preview",
-} as const;
-
-const GEMINI_IMAGE_MODEL_ENV_KEYS = {
-  flash: "GEMINI_FLASH_IMAGE_MODEL",
-  pro: "GEMINI_PRO_IMAGE_MODEL",
-} as const;
-
-type GeminiModelVariant = keyof typeof DEFAULT_GEMINI_IMAGE_MODELS;
 
 type ModelId =
   | "gemini-flash"
@@ -31,14 +23,6 @@ interface GenerateRequest {
   prompt: string;
   model: ModelId;
   size?: string;
-}
-
-export function resolveGeminiImageModel(
-  modelVariant: GeminiModelVariant,
-  env: NodeJS.ProcessEnv = process.env,
-): string {
-  const configuredModel = env[GEMINI_IMAGE_MODEL_ENV_KEYS[modelVariant]]?.trim();
-  return configuredModel || DEFAULT_GEMINI_IMAGE_MODELS[modelVariant];
 }
 
 async function generateWithGemini(
